Rework Cli.Mode.Bench.
Introduce a job dispatcher in addition to the workers. The dispatcher
produces jobs for workers to run. This will be used to make sure that
with a fixed seed, the set of games produced is independent of the
number of workers (and what they do). This wasn't the case
previously.